envsetup: Fix aosp & caf remote
Fixes: sed: -e expression #1, char 9: unknown option to `s'

* Return if .git directory is missing and move it on the top
* Update cmremote strings ( match caf/aosp remote )

+++ envsetup.sh
@@ -1788,13 +1788,13 @@ function addaokp() {
 
 function aokpremote()
 {
+    if ! git rev-parse --git-dir &> /dev/null
+    then
+        echo ".git directory not found. Please run this from the root directory of the Android repository you wish to set up."
+        return 1
+    fi
     git remote rm aokpremote 2> /dev/null
     GERRIT_REMOTE=$(git config --get remote.github.projectname)
-    if [ -z "$GERRIT_REMOTE" ]
-    then
-        echo Unable to set up the git remote, are you under a git repo?
-        return 0
-    fi
     AOKPUSER=$(git config --get review.gerrit.aokp.co.username)
     if [ -z "$AOKPUSER" ]
     then
@@ -1802,17 +1802,18 @@ function aokpremote()
     else
         git remote add aokpremote ssh://$AOKPUSER@gerrit.aokp.co:29418/$GERRIT_REMOTE
     fi
-    echo You can now push to "aokpremote".
+    echo "Remote 'aokpremote' created"
 }
 
 function aospremote()
 {
-    git remote rm aosp 2> /dev/null
-    if [ ! -d .git ]
+    if ! git rev-parse --git-dir &> /dev/null
     then
-        echo .git directory not found. Please run this from the root directory of the Android repository you wish to set up.
+        echo ".git directory not found. Please run this from the root directory of the Android repository you wish to set up."
+        return 1
     fi
-    PROJECT=`pwd -P | sed s#$ANDROID_BUILD_TOP/##g`
+    git remote rm aosp 2> /dev/null
+    PROJECT=${$(pwd -P)#$ANDROID_BUILD_TOP/}
     if (echo $PROJECT | grep -qv "^device")
     then
         PFX="platform/"
@@ -1823,12 +1824,13 @@ function aospremote()
 
 function cafremote()
 {
-    git remote rm caf 2> /dev/null
-    if [ ! -d .git ]
+    if ! git rev-parse --git-dir &> /dev/null
     then
-        echo .git directory not found. Please run this from the root directory of the Android repository you wish to set up.
+        echo ".git directory not found. Please run this from the root directory of the Android repository you wish to set up."
+        return 1
     fi
-    PROJECT=`pwd -P | sed s#$ANDROID_BUILD_TOP/##g`
+    git remote rm caf 2> /dev/null
+    PROJECT=${$(pwd -P)#$ANDROID_BUILD_TOP/}
     if (echo $PROJECT | grep -qv "^device")
     then
         PFX="platform/"
